+++ src/main/java/ee/laus/eventapp/payment/PaymentTypeRepository.java
@@ -2,5 +2,5 @@ package ee.laus.eventapp.payment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentTypeRepository extends JpaRepository<PaymentType, Integer> {
+public interface PaymentTypeRepository extends JpaRepository<PaymentType, Long> {
 }
